requirements.txt
pydantic>=2.5.0,<3.0.0
langgraph>=0.0.15
langchain-core>=0.1.0
python-dotenv
pyyaml
tavily-python
langchain_groq
langchain_community
chromadb
lxml
fastembed
unstructured
